[manuf] move provisioning util functions to separate lib
