+++ sw/host/provisioning/cp_lib/BUILD
@@ -21,9 +21,7 @@ rust_library(
         "//sw/host/opentitanlib",
         "//sw/host/provisioning/ujson_lib",
         "@crate_index//:anyhow",
-        "@crate_index//:arrayvec",
         "@crate_index//:clap",
-        "@crate_index//:hex",
         "@crate_index//:log",
         "@crate_index//:serde",
         "@crate_index//:serde_json",
