+++ sw/host/provisioning/cp_lib/src/lib.rs
@@ -5,9 +5,7 @@
 use std::time::Duration;
 
 use anyhow::{Context, Result};
-use arrayvec::ArrayVec;
 use clap::{ArgAction, Args};
-use hex::decode;
 
 use opentitanlib::app::TransportWrapper;
 use opentitanlib::dif::lc_ctrl::{DifLcCtrlState, DifLcCtrlToken, LcCtrlReg};
@@ -288,18 +286,4 @@ pub fn reset_and_lock(
     transport.pin_strapping("PINMUX_TAP_LC")?.remove()?;
 
     Ok(())
-}
-
-pub fn hex_string_to_u32_arrayvec<const N: usize>(hex_str: &str) -> Result<ArrayVec<u32, N>> {
-    let hex_str_no_sep = hex_str.replace('_', "");
-    let hex_str_prefix = "0x";
-    let sanitized_hex_str = if hex_str.starts_with(hex_str_prefix) {
-        hex_str_no_sep.strip_prefix(hex_str_prefix).unwrap()
-    } else {
-        hex_str_no_sep.as_str()
-    };
-    Ok(decode(sanitized_hex_str)?
-        .chunks(4)
-        .map(|bytes| u32::from_be_bytes(bytes.try_into().unwrap()))
-        .collect::<ArrayVec<u32, N>>())
 }
